Adjusting PR comments
