+++ Client/Ecosia/UI/NTP/News/NTPNewsCellViewModel.swift
@@ -10,7 +10,7 @@ protocol NTPNewsCellDelegate: AnyObject {
     func openSeeAllNews()
 }
 
-class NTPNewsCellViewModel {
+final class NTPNewsCellViewModel {
     private let news = News()
     private (set) var items = [NewsModel]()
     private let images = Images(.init(configuration: .ephemeral))
